+++ src/callbacks/postKnifeInitBabyFunctionMap.ts
@@ -1,11 +1,16 @@
+import { RandomBabyType } from "../babies";
+
 export const postKnifeInitBabyFunctionMap = new Map<
   int,
   (knife: EntityKnife) => void
 >();
 
 // Brother Bobby
-postKnifeInitBabyFunctionMap.set(522, (knife: EntityKnife) => {
-  // Make the knife invisible
-  knife.EntityCollisionClass = EntityCollisionClass.ENTCOLL_NONE;
-  knife.Visible = false;
-});
+postKnifeInitBabyFunctionMap.set(
+  RandomBabyType.BROTHER_BOBBY,
+  (knife: EntityKnife) => {
+    // Make the knife invisible
+    knife.EntityCollisionClass = EntityCollisionClass.ENTCOLL_NONE;
+    knife.Visible = false;
+  },
+);